requirement.txt
pandas
mysql
tqdm
pyyaml
mysql-connector-python